refactor: relocate CLI error helpers into cli package
Move cli-errors.js from src/shared to src/cli because the utilities are only consumed by CLI entrypoints; keeping them beside the CLI code clarifies ownership and avoids leaking CLI-specific logic into shared helpers.
