+++ src/cli/cli-errors.js
